+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG
 
  * Deprecate parameters `container.dumper.inline_factories` and `container.dumper.inline_class_loader`, use `.container.dumper.inline_factories` and `.container.dumper.inline_class_loader` instead
  * `FileProfilerStorage` removes profiles automatically after two days
- * Add `#[HttpStatus]` for defining status codes for exceptions
+ * Add `#[WithHttpStatus]` for defining status codes for exceptions
  * Use an instance of `Psr\Clock\ClockInterface` to generate the current date time in `DateTimeValueResolver`
  * Add `#[WithLogLevel]` for defining log levels for exceptions